paddle-ocr-server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
paddleocr==2.7.0
paddlepaddle==2.6.0
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
